#88 Extract the RowActionBuilder methods in another interface.
